Simplify sensors by not having it split in two, improve naming of variables Remove functionality to update ip address on new discovery
